+++ components/animations/index.ts
@@ -1,0 +1,4 @@
+export { CryptoBackgroundAnimation } from './CryptoBackgroundAnimation';
+export { TechBackgroundAnimation } from './TechBackgroundAnimation';
+export { EnhancedBackgroundAnimation } from './EnhancedBackgroundAnimation';
+export { PriceTickerAnimation } from './PriceTickerAnimation';